requirements.txt
asgiref==3.3.0
Django==3.1.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.3.1
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.4.1
psycopg2==2.8.6
django-leaflet==0.27.1